[MIG] website_geoengine_store_locator: Migration to 18.0
