+++ website_geoengine_store_locator/readme/CONFIGURE.md
@@ -1,12 +1,12 @@
 You should configure first a web site. Then drag n drop the Store
 Locator widget from the OCA/Geopspatial section.
 
-![](./static/description/snippet.png)
+![](../static/description/snippet.png)
 
 You can the configure the maximum number of results you want to display
 before to force your user to filter results.
 
-![](./static/description/snippet_configuration.png)
+![](../static/description/snippet_configuration.png)
 
 Then go to 'Contact' app and add some store type address to your
 contacts and add them some tags.
