+++ website_geoengine_store_locator/static/description/index.html
@@ -369,10 +369,10 @@ ul.auto-toc {
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! source digest: sha256:fe322224dbdd376eea03ba9aa37eaf3c4bc975d36a999b33a2f799caf36e441f
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/geospatial/tree/16.0/website_geoengine_store_locator"><img alt="OCA/geospatial" src="https://img.shields.io/badge/github-OCA%2Fgeospatial-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/geospatial-16-0/geospatial-16-0-website_geoengine_store_locator"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/geospatial&amp;target_branch=16.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
-<p>This module extends the <tt class="docutils literal">website_geoengine</tt> odoo module, to add
-a map snippet view for the website.
-By default the map features are filtered to only display contact of type <cite>store</cite>.</p>
+<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/geospatial/tree/18.0/website_geoengine_store_locator"><img alt="OCA/geospatial" src="https://img.shields.io/badge/github-OCA%2Fgeospatial-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/geospatial-18-0/geospatial-18-0-website_geoengine_store_locator"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/geospatial&amp;target_branch=18.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
+<p>This module extends the <tt class="docutils literal">website_geoengine</tt> odoo module, to add a map
+snippet view for the website. By default the map features are filtered
+to only display contact of type store.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -389,28 +389,25 @@ By default the map features are filtered to only display contact of type <cite>s
 </div>
 <div class="section" id="configuration">
 <h1><a class="toc-backref" href="#toc-entry-1">Configuration</a></h1>
-<p>You should configure first a web site. Then drag n drop the Store Locator widget from the OCA/Geopspatial section.</p>
-<div class="figure">
-<img alt="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/snippet.png" src="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/snippet.png" />
-</div>
-<p>You can the configure the maximum number of results you want to display before to force your user to filter results.</p>
-<div class="figure">
-<img alt="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/snippet_configuration.png" src="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/snippet_configuration.png" />
-</div>
-<p>Then go to ‘Contact’ app and add some store type address to your contacts and add them some tags.</p>
+<p>You should configure first a web site. Then drag n drop the Store
+Locator widget from the OCA/Geopspatial section.</p>
+<p><img alt="image1" src="https://raw.githubusercontent.com/OCA/geospatial/18.0/website_geoengine_store_locator/static/description/snippet.png" /></p>
+<p>You can the configure the maximum number of results you want to display
+before to force your user to filter results.</p>
+<p><img alt="image2" src="https://raw.githubusercontent.com/OCA/geospatial/18.0/website_geoengine_store_locator/static/description/snippet_configuration.png" /></p>
+<p>Then go to ‘Contact’ app and add some store type address to your
+contacts and add them some tags.</p>
 </div>
 <div class="section" id="usage">
 <h1><a class="toc-backref" href="#toc-entry-2">Usage</a></h1>
-<div class="figure">
-<img alt="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/screencast.gif" src="https://raw.githubusercontent.com/OCA/geospatial/16.0/website_geoengine_store_locator/static/description/screencast.gif" />
-</div>
+<p><img alt="screencast" src="https://raw.githubusercontent.com/OCA/geospatial/18.0/website_geoengine_store_locator/static/description/screencast.gif" /></p>
 </div>
 <div class="section" id="bug-tracker">
 <h1><a class="toc-backref" href="#toc-entry-3">Bug Tracker</a></h1>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/geospatial/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/geospatial/issues/new?body=module:%20website_geoengine_store_locator%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/geospatial/issues/new?body=module:%20website_geoengine_store_locator%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -439,7 +436,7 @@ mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
 <p>Current <a class="reference external" href="https://odoo-community.org/page/maintainer-role">maintainer</a>:</p>
 <p><a class="reference external image-reference" href="https://github.com/Wouitmil"><img alt="Wouitmil" src="https://github.com/Wouitmil.png?size=40px" /></a></p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/geospatial/tree/16.0/website_geoengine_store_locator">OCA/geospatial</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/geospatial/tree/18.0/website_geoengine_store_locator">OCA/geospatial</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
