+++ website_geoengine_store_locator/static/src/scss/snippets/s_openlayer_store_locator/frontend.scss
@@ -12,8 +12,60 @@ $s-map-desc-hover-alpha: 0.55 !default;
 
         .map {
             @include o-position-absolute(0, 0, 0, 0);
+            width: 100%;
+            height: 100%;
         }
     }
+
+    .search {
+        position: absolute;
+        top: 0.5rem;
+        left: 3rem;
+        min-width: 15rem;
+
+        input {
+            border-color: lightgrey;
+            border-width: 1px;
+            border-style: solid;
+            border-radius: 2px;
+
+            &:focus {
+                outline: none;
+            }
+        }
+
+        // Hide the old field
+        > .flexdatalist-alias.search-flexdatalist {
+            visibility: hidden;
+        }
+
+        .value {
+            margin-left: 0;
+            margin-right: 0;
+        }
+
+        .text {
+            background-color: lightgray;
+            border-radius: 5px;
+            border-color: gray;
+            border-width: 1px;
+            border-style: solid;
+            padding: 0.2rem 0.4rem;
+        }
+    }
+
+    .message {
+        position: absolute;
+        width: 20%;
+        z-index: 999;
+        top: 1rem;
+        left: 40%;
+        background: white;
+        border: rgb(211, 211, 211);
+        padding: 0.5rem;
+        text-align: center;
+    }
+
     .description {
         @include o-position-absolute(auto, 0, 0, 0);
         z-index: 99;
@@ -46,6 +98,95 @@ $s-map-desc-hover-alpha: 0.55 !default;
         @include o-position-absolute(0, 0, 0, 0);
         position: absolute !important;
         pointer-events: none;
+    }
+}
+
+.s_store_locator_popover {
+    min-width: 14rem;
+    padding: 0.65rem 0.75rem;
+    color: #383838;
+    line-height: 1.45;
+    animation: s_store_locator_popover_in 180ms ease-out both;
+
+    .s_store_locator_popover_header {
+        display: flex;
+        align-items: center;
+        gap: 0.5rem;
+        padding: 0.15rem 0.1rem 0.7rem;
+        margin-bottom: 0.7rem;
+        color: var(--o-color-1, #714b67);
+        border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+        font-size: 1rem;
+
+        .fa {
+            flex: 0 0 1rem;
+            text-align: center;
+            animation: s_store_locator_popover_icon_in 350ms ease-out both;
+        }
+    }
+
+    .s_store_locator_popover_address {
+        padding: 0 0.1rem 0.35rem 1.5rem;
+        margin-bottom: 0.35rem;
+    }
+
+    .s_store_locator_popover_meta {
+        display: flex;
+        align-items: flex-start;
+        gap: 0.5rem;
+        padding: 0.25rem 0.1rem;
+        margin-top: 0.2rem;
+        animation: s_store_locator_popover_row_in 220ms ease-out both;
+
+        &.s_store_locator_popover_meta_hours {
+            animation-delay: 50ms;
+        }
+
+        .fa {
+            flex: 0 0 1rem;
+            padding-top: 0.2rem;
+            color: var(--o-color-1, #714b67);
+            text-align: center;
+        }
+    }
+}
+
+@keyframes s_store_locator_popover_in {
+    from {
+        opacity: 0;
+        transform: translateY(0.4rem) scale(0.98);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0) scale(1);
+    }
+}
+
+@keyframes s_store_locator_popover_icon_in {
+    from {
+        transform: scale(0.7) rotate(-10deg);
+    }
+    to {
+        transform: scale(1) rotate(0);
+    }
+}
+
+@keyframes s_store_locator_popover_row_in {
+    from {
+        opacity: 0;
+        transform: translateX(-0.25rem);
+    }
+    to {
+        opacity: 1;
+        transform: translateX(0);
+    }
+}
+
+@media (prefers-reduced-motion: reduce) {
+    .s_store_locator_popover,
+    .s_store_locator_popover .fa,
+    .s_store_locator_popover .s_store_locator_popover_meta {
+        animation: none;
     }
 }
