--- website_geoengine_store_locator/static/styles.css
@@ -1,46 +1,0 @@
-.map {
-    width: 100%;
-    height: 100%;
-}
-.search {
-    position: absolute;
-    top: 0.5rem;
-    left: 3rem;
-    min-width: 15rem;
-}
-.search input {
-    border-color: lightgrey;
-    border-width: 1px;
-    border-style: solid;
-    border-radius: 2px;
-}
-.search input:focus {
-    outline: none;
-}
-/* Hide the old field */
-.search > .flexdatalist-alias.search-flexdatalist {
-    visibility: hidden;
-}
-.search .value {
-    margin-left: 0;
-    margin-right: 0;
-}
-.search .text {
-    background-color: lightgray;
-    border-radius: 5px;
-    border-color: gray;
-    border-width: 1px;
-    border-style: solid;
-    padding: 0.2rem 0.4rem;
-}
-.message {
-    position: absolute;
-    width: 20%;
-    z-index: 999;
-    top: 1rem;
-    left: 40%;
-    background: white;
-    border: rgb(211, 211, 211) solid 1px;
-    padding: 0.5rem;
-    text-align: center;
-}
